integration/docker: move CPU tests
To make tests structure clearer, move CPU test from run_test to cpu_test.
More CPU tests are coming.
